Added version with FLANN matcher and refactored stash directory
