+++ src/stash/face/FaceDetectionHaarCascade.py
