+++ src/stash/face/haarcascade_frontalface_alt2.xml
