+++ src/stash/media/panorama.png
